IAttackable Interface Created
IAttackable Interface Created - Used for Enemies attacking things

+++ Assets/Master/Assets/Interfaces/IDamageable.cs
@@ -7,5 +7,5 @@ public interface IDamageable
 {                                                      
     //event EventHandler OnMyEvent;       // can have events
     //int MyInt { get; set; }             // can have property, cannot have feilds 
-    void Damage(int baseDamageValue = 0);
+    void Damage(float baseDamageValue = 0);
 }
